backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
requests==2.32.3
beautifulsoup4==4.13.3
email-validator==2.2.0
PyJWT==2.9.0
